+++ packages/apps/dev-wallet/src/modules/key-source/key-source.repository.ts
@@ -42,8 +42,6 @@ export interface HDWalletRepository {
   getKeySource: (id: string) => Promise<KeySourceType>;
   addKeySource: (profile: KeySourceType) => Promise<void>;
   updateKeySource: (profile: KeySourceType) => Promise<void>;
-  getEncryptedValue: (key: string) => Promise<Uint8Array>;
-  addEncryptedValue: (key: string, value: string | Uint8Array) => Promise<void>;
   updateEncryptedValue: (
     key: string,
     value: string | Uint8Array,
@@ -64,15 +62,6 @@ const createKeySourceRepository = ({
     },
     updateKeySource: async (keySource: KeySourceType): Promise<void> => {
       return update('keySource', keySource);
-    },
-    getEncryptedValue: async (key: string): Promise<Uint8Array> => {
-      return getOne('encryptedValue', key);
-    },
-    addEncryptedValue: async (
-      key: string,
-      value: string | Uint8Array,
-    ): Promise<void> => {
-      return add('encryptedValue', value, key);
     },
     updateEncryptedValue: async (key: string, value: string | Uint8Array) => {
       return update('encryptedValue', value, key);