+++ packages/apps/dev-wallet/src/modules/transaction/transaction.service.ts
@@ -1,19 +1,31 @@
-import { IUnsignedCommand } from '@kadena/client';
-import { transactionRepository } from './transaction.repository';
+import { ChainId, IUnsignedCommand } from '@kadena/client';
+import { ITransaction, transactionRepository } from './transaction.repository';
 
-export async function addTransaction(
-  command: IUnsignedCommand,
-  profileId: string,
-  networkId: string,
-  groupId?: string,
-) {
-  const tx = {
+export async function addTransaction({
+  transaction,
+  profileId,
+  networkId,
+  groupId,
+  autoContinue = false,
+  crossChainId,
+}: {
+  transaction: IUnsignedCommand;
+  profileId: string;
+  networkId: string;
+  groupId?: string;
+  autoContinue?: boolean;
+  crossChainId?: ChainId;
+}) {
+  const tx: ITransaction = {
+    ...transaction,
     uuid: crypto.randomUUID(),
     profileId,
     networkId,
     status: 'initiated' as const,
     groupId,
-    ...command,
+    ...(crossChainId || autoContinue
+      ? { continuation: { crossChainId, autoContinue } }
+      : {}),
   };
   await transactionRepository.addTransaction(tx);
   return tx;