+++ packages/apps/dev-wallet/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @kadena/dev-wallet
 
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies \[78e6f88]
+- Updated dependencies \[eccd46e]
+- Updated dependencies \[eac8c84]
+  - @kadena/chainweb-node-client\@0.7.0
+  - @kadena/client\@1.15.0
+  - @kadena/client-utils\@0.11.0
+  - @kadena/pactjs-generator\@1.15.0
+
 ## 0.4.1
 
 ### Patch Changes